Implementing idle reason details for structures. Not as large of a change as it would appear but there were a lot of small changes to make this happen. Closes #89.

+++ src/States/MapViewState.cpp
@@ -796,7 +796,7 @@ void MapViewState::placeRobot()
 			mMineOperationsWindow.hide();
 			mTileMap->removeMineLocation(mTileMap->tileMouseHover());
 			tile->pushMine(nullptr);
-			for (size_t i = 0; i <= mTileMap->maxDepth(); ++i)
+			for (size_t i = 0; i <= static_cast<size_t>(mTileMap->maxDepth()); ++i)
 			{
 				Tile* _t = mTileMap->getTile(mTileMap->tileMouseHoverX(), mTileMap->tileMouseHoverY(), i);
